refactor: make `println` tests integration tests
