+++ tests/README.md
@@ -1,0 +1,11 @@
+# `println` tests
+
+Printed lines are displayed in `DebugLogger`, which is built on `env_logger`. The logger initialized by `env_logger` is global per process. To let each `println` test have its own process-local `DebugLogger`, we create a separate file here for each `println` test. The way integration tests are built and run ensures that this gives each `println` test its own `DebugLogger`. Otherwise multiple tests would concurrently write to the same `DebugLogger`.
+
+`DebugLogger` is feature-gated behind `tui`. To apply that gate to integration tests, add an entry according to the following pattern for each relevant file:
+
+```toml
+[[test]]
+name = "println_smoke_test"
+required-features = ["tui"]
+```
